refactor(codegen): split paragraphs/schemas/proofs into codegen/ submodules
Phase 1 Batch 1 (Moves 1-3 from docs/PHASE1-PLAN.md).  Extract 23
handler methods plus 7 helpers from latex_gen.py into three new
mixin modules under src/txt2tex/codegen/.

Moves are byte-identical Fowler Move-Method / Extract-Class
refactorings.  No method body, signature, name, or docstring was
modified.  LaTeXGenerator now composes the mixins via multiple
inheritance; runtime dispatch is unchanged because every handler
is reachable through the MRO.

* paragraphs.py — GivenType, FreeType, SyntaxBlock,
  SyntaxDefinition helper, Abbreviation, AxDef, GenDef, Zed.
* schemas.py — Schema, HorizDef, schema-text inline form,
  schema-calculus operators (rename, compose, pipe, hide, project),
  _emit_schema_inclusion declaration helper.
* proofs.py — ProofTree, InfruleBlock, ArgueChain
  (ARGUE/EQUIV/EQUAL), plus the proof-node internals
  (_generate_proof_node_infer, assumption scoping, case analysis),
  justification-label/text formatting, tree-depth calculation.

The TYPE_CHECKING block on CodegenDispatch declares cross-cutting
state and helper signatures so the mixin method bodies type-check
without modification.  pyproject.toml extends the existing ARG002
allowlist (singledispatch visitor parent arg) to codegen/*.py.

Gate 1 verified: make check passes (lint + format + mypy + pyright
+ 4700 unit tests); make test-e2e passes (159/159 byte-for-byte
fixtures).

latex_gen.py shrinks from 6864 → 4748 lines.

Refs: docs/PHASE1-PLAN.md, docs/DESIGN-ROADMAP.md.
