+++ src/txt2tex/codegen/__init__.py
@@ -2,4 +2,9 @@
 
 from __future__ import annotations
 
-from txt2tex.codegen import _smoke  # noqa: F401  # pyright: ignore[reportUnusedImport]
+from txt2tex.codegen import (
+    _smoke,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    paragraphs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    proofs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    schemas,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+)
